* Align preferences with config model store, removing preferences now moved to config model. * Document IPreferencesService * Document ITime
