+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/core/internal/CoreModule.kt
@@ -54,7 +54,7 @@ import com.onesignal.core.internal.outcomes.impl.OutcomeEventsRepository
 import com.onesignal.core.internal.permissions.IRequestPermissionService
 import com.onesignal.core.internal.permissions.impl.RequestPermissionService
 import com.onesignal.core.internal.preferences.IPreferencesService
-import com.onesignal.core.internal.preferences.PreferencesService
+import com.onesignal.core.internal.preferences.impl.PreferencesService
 import com.onesignal.core.internal.purchases.TrackAmazonPurchase
 import com.onesignal.core.internal.purchases.TrackGooglePurchase
 import com.onesignal.core.internal.service.ServiceBuilder
@@ -65,7 +65,7 @@ import com.onesignal.core.internal.startup.IBootstrapService
 import com.onesignal.core.internal.startup.IStartableService
 import com.onesignal.core.internal.startup.StartupService
 import com.onesignal.core.internal.time.ITime
-import com.onesignal.core.internal.time.Time
+import com.onesignal.core.internal.time.impl.Time
 import com.onesignal.core.internal.user.ISubscriptionManager
 import com.onesignal.core.internal.user.SubscriptionManager
 import com.onesignal.core.internal.user.UserManager
