+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/core/internal/time/ITime.kt
@@ -1,6 +1,13 @@
 package com.onesignal.core.internal.time
 
+/**
+ * Provides an abstraction to retrieving the current time.  This should be used rather
+ * than standard library services, to allow for more flexible testing scenarios.
+ */
 internal interface ITime {
+    /**
+     * Returns the current time in unix time milliseconds (the number of milliseconds between the
+     * current time and midnight, January 1, 1970 UTC).
+     */
     val currentTimeMillis: Long
-    val elapsedRealtime: Long
 }
