--- OneSignalSDK/onesignal/src/main/java/com/onesignal/core/internal/time/Time.kt
@@ -1,10 +1,0 @@
-package com.onesignal.core.internal.time
-
-import android.os.SystemClock
-
-internal class Time : ITime {
-    override val currentTimeMillis: Long
-        get() = System.currentTimeMillis()
-    override val elapsedRealtime: Long
-        get() = SystemClock.elapsedRealtime()
-}
