+++ OneSignalSDK/onesignal/src/main/java/com/onesignal/core/internal/time/impl/Time.kt
@@ -1,0 +1,8 @@
+package com.onesignal.core.internal.time.impl
+
+import com.onesignal.core.internal.time.ITime
+
+internal class Time : ITime {
+    override val currentTimeMillis: Long
+        get() = System.currentTimeMillis()
+}
